refactor(apps): group app routes + tests into the app/ entity folder
Move routes/app.ts -> routes/app/app.routes.ts and split the monolithic
routes/app.test.ts into one test file per endpoint (create/get/list/update/
delete/versions/tools/diagnostics/screenshot/templates/external + a kept
lifecycle test), matching the virtual-api-key/skill entity-folder convention.

The routes file is a content-identical move (only its path and the
routes/index.ts import change); no runtime behavior changes. Per-endpoint
coverage grows: 404/403 authorization paths per action and the previously
untested single-version GET /api/apps/:appId/versions/:version.
