+++ platform/backend/src/routes/index.ts
@@ -4,7 +4,7 @@ export { default as a2aV2Routes } from "./a2a-v2";
 export { default as agentRoutes } from "./agent";
 export { default as agentToolRoutes } from "./agent-tool";
 export { default as apiKeyRoutes } from "./api-key";
-export { default as appRoutes } from "./app";
+export { default as appRoutes } from "./app/app.routes";
 export { default as archestraCatalogProxyRoutes } from "./archestra-catalog-proxy";
 export { default as auditLogRoutes } from "./audit-log";
 export { default as authRoutes } from "./auth";
